requirements.txt
Django==4.1.0
djangorestframework==3.14.0
mysqlclient==2.2.1